requirements.txt
discord.py==1.4.1
requests==2.22.0
beautifulsoup4==4.7.1
lxml==4.5.0
matplotlib==3.0.3